[UR][TEST] Signal IPC event via reusable-events API instead of Level Zero
The IPC event conformance fixture signaled the producer event by driving the
Level Zero API directly. Now that urEventCreateExp produces a counter-based
reusable event, the fixture can signal it through the reusable-events API instead.
